companionbot-backend/requirements.txt
flask
flask-cors
requests
scikit-learn
pymongo
python-dotenv
